Refactor
This commit is a full refactor of the base of the plugin/provider
as well as the openstack_compute resource.

At the moment, it's able to launch and destroy an instance using
the latest version of gophercloud. It might be able to do more,
but I haven't fully tested it.

Some functionality has been removed, such as haklop's network
extensions and floating IPs. It shouldn't be very hard to add this
stuff back in, though.