Network updates
This commit renames the `network_advanced` parameter to just
`network`.

It also adds the `network_info` parameter which will return some
basic network information about each network that an instance is
configured with.

Finally, a parameter called `network_provider` is added for future
use of working with either nova-network or Neutron.