API and Volume attach refactor
This commit moves the recently added api_version from the resources
to the provider. This is because there shouldn't be a user-case where
the user would launch different resources in the same cloud using
different api versions.

This commit also refactors volume attachment support. Previously,
the volume was attached on the volume resource, but it's now attached
by the instance. However, the volume still needs to know about the
attachment both for reporting purposes as well as safety of deletion.